feat: refactor dashboard components by removing old implementations and adding new styled and config files for BottomNavBar and DashboardCard
