+++ src/presentation/widgets/dashboard/BottomNavBar/BottomNavBar.styled.ts
@@ -1,0 +1,35 @@
+import styled from "styled-components";
+
+const NavBar = styled.nav`
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  height: 62px;
+  width: 100vw;
+  padding: 0 20px;
+  background-color: var(--background);
+  border-top: 1px solid #cbcbcb;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  z-index: 999;
+`;
+
+const NavItem = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  align-items: center;
+  justify-content: center;
+  width: 70px;
+  height: 38px;
+`;
+
+const Label = styled.span`
+  font-size: 10px;
+  font-weight: 500;
+  letter-spacing: -0.24px;
+  color: var(--txt-grey);
+  text-align: center;
+`;
+export { NavBar, NavItem, Label };
