+++ src/presentation/widgets/dashboard/BottomNavBar/index.ts
@@ -1,0 +1,1 @@
+export { BottomNavBar } from "./BottomNavBar";
