+++ src/presentation/widgets/dashboard/DashboardCard/index.ts
@@ -1,0 +1,2 @@
+export { DashboardCard } from "./DashboardCard";
+export { mockDashboardCardsData } from "./DashboardCard.mocks";
